+++ c/libmcpb/include/mcpb/mcpb.h
@@ -44,8 +44,8 @@ const char *mcpb_strerror(int err);
 
 #define MCPB_VERSION_MAJOR 0
 #define MCPB_VERSION_MINOR 2
-#define MCPB_VERSION_PATCH 0
-#define MCPB_VERSION_STRING "0.2.0"
+#define MCPB_VERSION_PATCH 1
+#define MCPB_VERSION_STRING "0.2.1"
 
 #ifdef __cplusplus
 } /* extern "C" */